add embeddables, move compositeField, more stable field_name handling
